Move gematria tests to a file for each method
